Refactor code to many other files
