+++ src/main/java/FormatPrinter.java
@@ -1,0 +1,5 @@
+public class FormatPrinter {
+    public static void print(String str) {
+                System.out.println(Formatter.format_response(str));
+    }
+}
