+++ src/main/java/Formatter.java
@@ -1,0 +1,8 @@
+public class Formatter {
+    public static String format_response(String output_msg) {
+        return 
+        "____________________________________________________________\n"+
+        output_msg +
+        "____________________________________________________________\n";
+    }
+}
