+++ src/main/java/NoDescriptionException.java
@@ -1,5 +1,5 @@
 public class NoDescriptionException extends Exception {
     public NoDescriptionException(String message) {
-        super(message);
+        super(message + "\n");
     }
 }
