+++ src/main/java/Ui.java
@@ -1,0 +1,19 @@
+import java.util.Scanner;
+
+public class Ui {
+    /**
+     * get_input() is in charge of asking for input from user
+     * and return the input as a string to be used for something else.
+     * @param sc Scanner object
+     * @return a string representing user input
+     */
+    public static String get_input(Scanner sc) {
+        String user_input = sc.nextLine().trim(); // Read user input
+        return user_input;
+    }
+
+
+
+
+
+}
